Expose `IndexManager` as an interface
This is consistent with how we expose `Database`
