+++ cloudant-sync-datastore-core/src/main/java/com/cloudant/sync/datastore/CloudantSync.java
@@ -2,6 +2,7 @@ package com.cloudant.sync.datastore;
 
 import com.cloudant.sync.datastore.encryption.KeyProvider;
 import com.cloudant.sync.query.IndexManager;
+import com.cloudant.sync.query.IndexManagerImpl;
 
 import java.io.IOException;
 import java.sql.SQLException;
@@ -17,7 +18,7 @@ public class CloudantSync {
 
     protected CloudantSync(String dir, String name, KeyProvider keyProvider) throws DatastoreException, IOException, SQLException {
         this.database = new DatabaseImpl(dir, name, keyProvider);
-        this.query = new IndexManager(database);
+        this.query = new IndexManagerImpl(database);
     }
 
 }
