+++ cloudant-sync-datastore-core/src/main/java/com/cloudant/sync/query/QueryConstants.java
@@ -52,7 +52,7 @@ final class QueryConstants {
 
     public static String[] getSchemaVersion1() {
         return new String[] {
-                "CREATE TABLE " + IndexManager.INDEX_METADATA_TABLE_NAME + " ( " +
+                "CREATE TABLE " + IndexManagerImpl.INDEX_METADATA_TABLE_NAME + " ( " +
                 "        index_name TEXT NOT NULL, " +
                 "        index_type TEXT NOT NULL, " +
                 "        field_name TEXT NOT NULL, " +
@@ -62,7 +62,7 @@ final class QueryConstants {
 
     public static String[] getSchemaVersion2() {
         return new String[] {
-                "ALTER TABLE " + IndexManager.INDEX_METADATA_TABLE_NAME +
+                "ALTER TABLE " + IndexManagerImpl.INDEX_METADATA_TABLE_NAME +
                 "        ADD COLUMN index_settings TEXT NULL;"
         };
     }
