+++ cloudant-sync-datastore-core/src/main/java/com/cloudant/sync/query/QuerySqlTranslator.java
@@ -141,7 +141,7 @@ class QuerySqlTranslator {
             String allDocsIndex = chooseIndexForFields(neededFields, indexes);
 
             if (allDocsIndex != null && !allDocsIndex.isEmpty()) {
-                String tableName = IndexManager.tableNameForIndex(allDocsIndex);
+                String tableName = IndexManagerImpl.tableNameForIndex(allDocsIndex);
                 String sql = String.format(Locale.ENGLISH, "SELECT _id FROM \"%s\"", tableName);
                 sqlNode.sql = SqlParts.partsForSql(sql, new String[]{});
             }
@@ -453,7 +453,7 @@ class QuerySqlTranslator {
             return null;
         }
 
-        String tableName = IndexManager.tableNameForIndex(indexName);
+        String tableName = IndexManagerImpl.tableNameForIndex(indexName);
 
         String sql = String.format(Locale.ENGLISH,
                                    "SELECT _id FROM \"%s\" WHERE %s",
@@ -480,7 +480,7 @@ class QuerySqlTranslator {
         Map<String, Object> textClause = (Map<String, Object>) clause;
         Map<String, String> searchClause = (Map<String, String>) textClause.get(TEXT);
 
-        String tableName = IndexManager.tableNameForIndex(indexName);
+        String tableName = IndexManagerImpl.tableNameForIndex(indexName);
         String search = searchClause.get(SEARCH);
         search = search.replace("'", "''");
 
@@ -563,7 +563,7 @@ class QuerySqlTranslator {
                 } else {
                     String whereClause;
                     String sqlOperator = operatorMap.get(operator);
-                    String tableName = IndexManager.tableNameForIndex(indexName);
+                    String tableName = IndexManagerImpl.tableNameForIndex(indexName);
                     String placeholder;
                     if (operator.equals(IN)) {
                         // The predicate map value must be a List here.
