+++ cloudant-sync-datastore-core/src/test/java/com/cloudant/sync/query/MockSQLOnlyIndexManager.java
@@ -19,15 +19,15 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * This sub class of the {@link com.cloudant.sync.query.IndexManager} along with
+ * This sub class of the {@link IndexManagerImpl} along with
  * {@link com.cloudant.sync.query.MockSQLOnlyQueryExecutor} is used by query
  * executor tests to force the tests to exclusively exercise the SQL engine logic.
  * This class is used for testing purposes only.
  *
- * @see com.cloudant.sync.query.IndexManager
+ * @see IndexManagerImpl
  * @see com.cloudant.sync.query.MockSQLOnlyQueryExecutor
  */
-public class MockSQLOnlyIndexManager extends IndexManager {
+public class MockSQLOnlyIndexManager extends IndexManagerImpl {
 
     public MockSQLOnlyIndexManager(Database database) {
         super(database);
