+++ cloudant-sync-datastore-core/src/test/java/com/cloudant/sync/replication/PullStrategyTest.java
@@ -27,7 +27,7 @@ import com.cloudant.mazha.Response;
 import com.cloudant.sync.datastore.DatabaseImpl;
 import com.cloudant.sync.datastore.DocumentRevision;
 import com.cloudant.sync.datastore.DocumentRevisionTree;
-import com.cloudant.sync.query.IndexManager;
+import com.cloudant.sync.query.IndexManagerImpl;
 import com.cloudant.sync.query.QueryResult;
 
 import org.junit.Assert;
@@ -386,7 +386,7 @@ public class PullStrategyTest extends ReplicationTestBase {
 
         Assert.assertEquals(0, datastore.getDocumentCount());
 
-        IndexManager im = new IndexManager(datastore);
+        IndexManagerImpl im = new IndexManagerImpl(datastore);
         try {
             im.ensureIndexed(Arrays.<Object>asList("diet"), "diet");
